+++ src/include/definitions.h
@@ -51,7 +51,7 @@
 // MEMORY USAGE: (units, structures, etc , more info below)
 #define MAX_UNITS        300     // max of units in the game
 #define MAX_STRUCTURES   200     // max of structures in the game
-#define MAX_BULLETS      300     // max of bullets in the game at once
+// #define MAX_BULLETS      300     // max of bullets in the game at once
 // #define MAX_PARTICLES    600	 // max of particles in the game at once
 #define MAX_SKIRMISHMAPS 100     // max of 100 skirmish maps
 
@@ -83,7 +83,7 @@
 #define MAX_STRUCTURETYPES  50    // max of building types
 
 // BULLETS stuff
-#define MAX_BULLET_TYPES      30
+//#define MAX_BULLET_TYPES      30
 
 // PARTICLES
 //#define MAX_PARTICLE_TYPES    50
